BoardConfig: Add system_ext to the list of partitions
https://github.com/sonyxperiadev/device-sony-lena/commit/8cf42709812ac470940d90059f3b66f1fef1e6eb
adds the system_ext partition to our build.  However, this partition
doesn't exist in our partition table yet, resulting in:

    Sending 'system_ext' (110876 KB)                   OKAY [  3.034s]
    Writing 'system_ext'                               FAILED (remote: 'No such file or directory')

When rewriting partition tables using fastboot-flashall.  After this
change the missing partition is properly taken into account:

    Resizing 'system_ext_a'                            OKAY [  0.005s]

Note that this partition is _not mandatory_, and should definitely not
be added "just because stock has it too".  This is a SUPER device
meaning we can configure the partition layout however we like.  On all
other devices the contents simply reside inside /system/system_ext.

Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>